src/functions/image-download-handler/requirements.txt
boto3==1.24.32
boto3-stubs==1.24.38
boto3-stubs-lite==1.24.38
requests==2.28.1
pillow==9.2.0
pydantic==1.9.1